requirements.txt
beautifulsoup4==4.12.3
blinker==1.7.0
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs
cryptography==41.0.7
Deprecated==1.2.14
deprecation==2.1.0
dnspython==2.4.2
dsinternals==1.2.4
exceptiongroup==1.2.0
Flask==3.0.0
flask-sock==0.7.0
future==0.18.3
gunicorn==21.2.0
h11==0.14.0
humanfriendly==10.0
img2pdf==0.5.1
impacket==0.11.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
ldap3==2.9.1
ldapdomaindump==0.9.4
lxml
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
ocrmypdf
packaging==23.2
pdfminer.six==20231228
pexpect==4.9.0
pikepdf==8.11.2
pillow==10.2.0
pluggy==1.3.0
ptyprocess==0.7.0
pyasn1==0.5.1
pycparser==2.21
pycryptodomex==3.19.1
Pygments==2.17.2
PyMuPDF==1.23.16
PyMuPDFb==1.23.9
pyOpenSSL==23.3.0
PyPDF2==3.0.1
pyspnego==0.10.2
pytest==7.4.4
rich==13.7.0
simple-websocket==1.0.0
six==1.16.0
soupsieve==2.5
tomli==2.0.1
tqdm==4.66.1
watchdog==3.0.0
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
wheel
